+++ upload/admin/controller/extension/payment/securetrading_ws.php
@@ -453,7 +453,7 @@ class ControllerExtensionPaymentSecureTradingWs extends Controller {
 					$json['data']['created'] = date("Y-m-d H:i:s");
 					$json['data']['amount'] = $amount;
 					$json['data']['release_status'] = $release_status;
-					$json['data']['total'] = (double)$total_released;
+					$json['data']['total'] = (float)$total_released;
 					$json['error'] = false;
 				}
 			} else {
@@ -519,8 +519,8 @@ class ControllerExtensionPaymentSecureTradingWs extends Controller {
 					$json['data'] = array();
 					$json['data']['created'] = date("Y-m-d H:i:s");
 					$json['data']['amount'] = $amount * -1;
-					$json['data']['total_released'] = (double)$total_released;
-					$json['data']['total_rebated'] = (double)$total_rebated;
+					$json['data']['total_released'] = (float)$total_released;
+					$json['data']['total_rebated'] = (float)$total_rebated;
 					$json['data']['rebate_status'] = $rebate_status;
 					$json['error'] = false;
 				} else {
